refactor: extract logic with contract title into a sep function
