+++ packages/shared/src/utils/helpers/dtrader-v2-helpers.ts
@@ -1,6 +1,9 @@
 /* TODO: remove this component after /trader package is separated into its own repo.
 It's used to keep dtrader_v2 utils that are currently shared between various packages. */
 
+import { extractInfoFromShortcode, isHighLow } from '../shortcode';
+import { getMarketName, getTradeTypeName } from './market-underlying';
+
 export const POSITIONS_V2_TAB_NAME = {
     OPEN: 'Open',
     CLOSED: 'Closed',
@@ -17,3 +20,18 @@ export const getPositionsV2TabIndexFromURL = () => {
 
 export const isDTraderV2 = () =>
     !!JSON.parse(localStorage.getItem('FeatureFlagsStore') ?? '{}')?.data?.dtrader_v2 && window.innerWidth < 600;
+
+export const getTradeNotificationMessage = (shortcode: string) => {
+    const extracted_info_from_shortcode = extractInfoFromShortcode(shortcode);
+    const symbol = getMarketName(extracted_info_from_shortcode.underlying);
+    const trade_type = extracted_info_from_shortcode.category;
+    const contract_type = getTradeTypeName(trade_type, {
+        isHighLow: isHighLow({ shortcode }),
+        showMainTitle: true,
+    });
+    const contract_type_with_subtype = `${contract_type} ${getTradeTypeName(trade_type, {
+        isHighLow: isHighLow({ shortcode }),
+    })}`.trim();
+
+    return `${contract_type_with_subtype} - ${symbol}`;
+};
